refactored patchdataset out so that it can be used seperately for research.
